idea: protobuf serializer
In our context, we use protobufs heavily, and probably will continue
to do so for the foreseeable future.

I believe it would be beneficial to have a ProtoSerializer so we can
leverage the other bits that exist here already, but still be able to
utilize protos as your primary models.